Added Files for Proj1 Reproduction
If the code works properly you should only need the datasets that I download in the RMD file, which are E0 and Database. The downloads need to be changed from my path files (ex: "Users//Windows//etc etc") to your own path files, and download all necessary libraries. Besides that you should be good to go